FastAPI/pyproject.toml
[tool.poetry]
name = "api"
version = "0.0.1"
description = ""
authors = ["sergio-fogel <fogel.sergio@gmail.com>"]
readme = "README.md"
packages = [{include = "FastAPI"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.89.1"}
pydantic = "^1.10.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
